Added item file the new item pickax and new state
Taking the idea of states, the new file items was created to hold in
information for any items that need or want to be created that influence
the miner or the states. For now there is only pickax which can
influence the miner's strength and capability of finding gold. To do
this the miner has defined build when created which defines his health
and strength along with a pickax variable that must be defined during
the miner's creaion. Finally the new shopping state was created which
can be entered when he as the bank depositing gold. While he is
despositiong gold the if statements will check his current savings and
his current pickax's strength to see if he can buy a new one. If he can
then he will leave the bank and go to the store and buy a new one which
will then subtract gold from his savings. From the shop he will go
directly back the mines.