Make plant death work
Made plants die properly at the end of their lifespan.

Corrected plant spawning conditions so plants would spawn

Removed the buff and debuff methods on Plant subclasses and replaced them with the effect method so that all plants have a method with the same name to alter entities with.  Currently none of the characters in the game are interacting with plants...

Some lingering PEP 8 cleanup from the last commit