Created the World
Reorganized the entities into an entities package.  The number of entities was getting too large to include in the main west_world.py file.  New entities can be created in their own module now and imported through the entities package.  Some thought will need to be given to how the states and items are organized in this project.  The states and items will most likely end up being structured in a similar fashion to the entities.  The biggest question I have right now is , "Should the items be entities?".

Created a new World object to hold all of the game entities.  This will allow entities to easily remove themselves from the game world when needed.  The World object also has an update method that replaces the update for loop over the game_objects list in west_world.py.  Creating the World object required modifying all of the entities to take world as an argument and to set it on the BaseGameEntity.

Random PEP 8 cleanup

Remove prints from the plant spawning process as they were not using the miner in the format.