Create StateMachine
Created the StateMachine based on Mat Buckland's StateMachine class on pages 64 and 65.  This class allows for a global state and a previous state so the entities can have more complex behaviors.  An added benefit to the StateMachine class is that it collects all of the state management code into one place so methods like change_state are not implemented on each entity.