Create messaging system
Implement the event driven messaging system detailed in Buckland on pages 69-83.  The messaging system allows for entities to interact with each other.  Buckland's example of interaction was the miner notifying his wife when he comes home which cuases her to begin cooking a stew.  Implementing the system required adding several components:

The Telegram object which contains a message, sender, receiver, dispatch time and allows for passing any other keyword data through kwargs.  Kwargs makes Buckland's ExtraInfo attribute unneeded.

Modification state machine to allow for message handling.

Global states for the Miner and Wife that allow for the current state to be interrupted when a message is received.

Edits to the Wife's states to more closely follow Buckland's example.

The MessageDispatcher to handle dispatching messages (instantaneous and delayed messages).  It feels like the MessageDispatcher should simply be combined into our World object in a later commit.